separated validation related func to a file
