+++ abs_summarization/main.py
@@ -2,6 +2,7 @@ import numpy as np
 import pandas as pd
 import torch
 import training
+import validation
 from custom_dataset import create_dataset
 from init_wandb import Wandb_Init
 from constants import Directories
@@ -20,51 +21,6 @@ https://colab.research.google.com/github/abhimishra91/transformers-tutorials/blo
 device = "cuda" if cuda.is_available() else "cpu"
 
 
-def validate(
-    epoch: int,
-    tokenizer: PreTrainedTokenizerBase,
-    model: tuple,
-    device: str,
-    loader: DataLoader,
-) -> tuple[list[str], list[str]]:
-    model.eval()
-    predictions = []
-    actuals = []
-    with torch.no_grad():
-        for _, data in enumerate(loader, 0):
-            y = data["target_ids"].to(device, dtype=torch.long)
-            ids = data["source_ids"].to(device, dtype=torch.long)
-            mask = data["source_mask"].to(device, dtype=torch.long)
-
-            generated_ids = model.generate(
-                input_ids=ids,
-                attention_mask=mask,
-                max_length=150,
-                num_beams=2,
-                repetition_penalty=2.5,
-                length_penalty=1.0,
-                early_stopping=True,
-            )
-            preds = [
-                tokenizer.decode(
-                    g, skip_special_tokens=True, clean_up_tokenization_spaces=True
-                )
-                for g in generated_ids
-            ]
-            target = [
-                tokenizer.decode(
-                    t, skip_special_tokens=True, clean_up_tokenization_spaces=True
-                )
-                for t in y
-            ]
-            if _ % 100 == 0:
-                print(f"Completed {_}")
-
-            predictions.extend(preds)
-            actuals.extend(target)
-    return predictions, actuals
-
-
 def read_training_dataset():
     df = pd.read_csv(
         Directories.TRAIN_SETS.joinpath("news_summary.csv"), encoding="latin-1"
@@ -73,26 +29,6 @@ def read_training_dataset():
     df.ctext = "summarize: " + df.ctext
     print(df.head())
     return df
-
-
-def start_validation(
-    wandb_init,
-    epoch: int,
-    tokenizer: PreTrainedTokenizerBase,
-    model: tuple,
-    device: str,
-    val_loader: DataLoader,
-):
-    print(
-        "Now generating summaries on our fine tuned model for the validation dataset and saving it in a dataframe"
-    )
-    for epoch in range(wandb_init._config.VAL_EPOCHS):
-        predictions, actuals = validate(epoch, tokenizer, model, device, val_loader)
-        final_df = pd.DataFrame({"Generated Text": predictions, "Actual Text": actuals})
-        final_df.to_csv(
-            "./models/predictions.csv"
-        )  # Saving the dataframe as predictions.csv
-        print("Output Files generated for review")
 
 
 def save_model(model_output: str, model: tuple, save: bool = True) -> None:
@@ -140,7 +76,7 @@ def main(which_llm: str, model_output: str, train_epoch: int = 2):
         wandb_init, tokenizer, model, device, training_loader, optimizer
     )
     save_model(model_output, model, True)
-    start_validation(
+    validation.start_validation(
         wandb_init, wandb_init._config.VAL_EPOCHS, tokenizer, model, device, val_loader
     )
